separate Results and play in two components
